Formatted documents.
Added currentOrder.
Added logic for sortOptionSelect().

Added a watcher for sortOption.
This is to detect that if the sortOption changes,
change currentOrder back to "asc".
This is to ensure that it remains Descending when switching sortOptions.

Hacky solution,
cause for some reason 'lessons' WILL NOT REMEMBER THAT IT WAS REVERSED.
So I worked around it.

Changed lessons.json to give different spaces to lessons.
Will change their locations as well.