Added Bootstrap Icons import.
Edited navbar stuff;
to make the "Products" and "Cart" buttons look nicer in mobile view.

Adjusted cartContents() such that I attached a reference to the lesson;
rather than decomposing it.
=> table v-for was altered to accomodate the change.

Added 'plus' and 'minus' btns to each of the cart entries;
to increase and decrease quantity of that particular lesson.
'plus' btn disappears when spaces of that lesson = 0.

Added currency units to cart table.

Renamed following variables:
cartEntry -> cartTableEntry
cartContents -> cartTableContents
cartEntries -> cartTableEntries
Did this change to reinforce the job of cartTableContents():
to provide the data needed for display in the cart page.
Moreover, it also allows it to have an additional responsibility:
to switch back to products page, if cart becomes empty on cart page.

Maybe I could just do away with cartTableContents(),
and have the 'cart' hold everything I need for display,
and use it directly?
This means that addTocart(), removeFromCart(), and cartPrice()
will need editing, but then that would be it.
I won't need cartTableContents() then.